# pkg/server/api/docs/swagger.yaml
components:
  schemas:
...
    config.Config:
      properties:
        boot_modules:
          items:
            type: string
          type: array
          uniqueItems: false
        dhcp:
          $ref: '#/components/schemas/config.DHCP'
        dns:
          $ref: '#/components/schemas/config.DNS'
        friends:
          items:
            $ref: '#/components/schemas/config.Friend'
          type: array
          uniqueItems: false
        gai:
          $ref: '#/components/schemas/config.GAIConfig'
        ha:
          $ref: '#/components/schemas/config.HA'
        hostname:
          type: string
        interfaces:
          additionalProperties:
            $ref: '#/components/schemas/config.Interface'
          type: object
        logging:
          $ref: '#/components/schemas/config.Logging'
        monitoring:
          $ref: '#/components/schemas/config.MonitoringConfig'
        nftables:
          $ref: '#/components/schemas/config.NftablesConfig'
        routing:
          $ref: '#/components/schemas/config.Routing'
        services:
          additionalProperties:
            $ref: '#/components/schemas/config.Service'
          type: object
        ssh:
          $ref: '#/components/schemas/config.SSH'
        sysctl:
          additionalProperties:
            type: string
          type: object
        tunnels:
          additionalProperties:
            $ref: '#/components/schemas/config.Tunnel'
          type: object
        users:
          additionalProperties:
            $ref: '#/components/schemas/config.User'
          type: object
        version:
          type: string
        vrfs:
          additionalProperties:
            $ref: '#/components/schemas/config.VRFConfig'
          type: object
        wireguard:
          additionalProperties:
            $ref: '#/components/schemas/config.Wireguard'
          type: object
      required:
      - hostname
      - version
      type: object
...
    config.DHCP:
      properties:
        control_agent:
          $ref: '#/components/schemas/config.KeaControlAgent'
        ddns:
          $ref: '#/components/schemas/config.DHCPDDNS'
        enabled:
          type: boolean
        interfaces:
          items:
            type: string
          type: array
          uniqueItems: false
        subnets4:
          items:
            $ref: '#/components/schemas/config.KeaSubnet'
          type: array
          uniqueItems: false
        subnets6:
          items:
            $ref: '#/components/schemas/config.KeaSubnet'
          type: array
          uniqueItems: false
      type: object
...
    config.DNS:
      properties:
        nameservers:
          items:
            type: string
          type: array
          uniqueItems: false
        search:
          items:
            type: string
          type: array
          uniqueItems: false
        server:
          $ref: '#/components/schemas/config.DNSServer'
      type: object
...
    config.Friend:
      properties:
        enabled:
          type: boolean
        ha:
          $ref: '#/components/schemas/config.FriendHA'
        hostname:
          type: string
        identity:
          $ref: '#/components/schemas/config.FriendIdentity'
        manage:
          type: boolean
        name:
          type: string
        sync:
          $ref: '#/components/schemas/config.FriendSync'
        tls_skip_verify:
          type: boolean
        token:
          type: string
        url:
          type: string
      required:
      - name
      - token
      - url
      type: object
...
    config.GAIConfig:
      properties:
        label:
          items:
            $ref: '#/components/schemas/config.GAIEntry'
          type: array
          uniqueItems: false
        precedence:
          items:
            $ref: '#/components/schemas/config.GAIEntry'
          type: array
          uniqueItems: false
      type: object
...
    config.HA:
      properties:
        conntrackd:
          $ref: '#/components/schemas/config.Conntrackd'
        vrrp:
          items:
            $ref: '#/components/schemas/config.VRRPInstance'
          type: array
          uniqueItems: false
      type: object
    config.Interface:
      properties:
        addresses:
          items:
            type: string
          type: array
          uniqueItems: false
        bond:
          $ref: '#/components/schemas/config.Bond'
        bridge:
          $ref: '#/components/schemas/config.Bridge'
        dhcp_options:
          items:
            type: string
          type: array
          uniqueItems: false
        mtu:
          type: integer
        select:
          type: string
        type:
          type: string
        vlan:
          $ref: '#/components/schemas/config.VLAN'
        vrf:
          type: string
        vxlan:
          $ref: '#/components/schemas/config.VXLAN'
      type: object
...
    config.Logging:
      properties:
        file:
          type: string
        level:
          type: string
        target:
          type: string
      type: object
...
    config.MonitoringConfig:
      properties:
        collection:
          $ref: '#/components/schemas/config.CollectionIntervals'
        lldp:
          $ref: '#/components/schemas/config.LLDPConfig'
        probes:
          items:
            $ref: '#/components/schemas/config.PingProbe'
          type: array
          uniqueItems: false
      type: object
...
    config.NftablesConfig:
      properties:
        chains:
          additionalProperties:
            $ref: '#/components/schemas/config.NftChain'
          type: object
        defines:
          type: string
        include:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
    config.Service:
      properties:
        after:
          type: string
        configs:
          items:
            $ref: '#/components/schemas/config.ServiceConfig'
          type: array
          uniqueItems: false
        enable:
          type: boolean
      required:
      - enable
      type: object
...
    config.VLAN:
      properties:
        id:
          type: integer
      required:
      - id
      type: object
    config.VRFConfig:
      properties:
        table:
          type: integer
      required:
      - table
      type: object
...
    types.APIKey:
      properties:
        created_at:
          type: integer
        created_by:
          type: string
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
        revoked_at:
          type: integer
      required:
      - created_at
      - created_by
      - id
      - name
      - prefix
      - revoked_at
      type: object
...
    types.ApplyResult:
      properties:
        bundleID:
          type: string
        errors:
          items:
            $ref: '#/components/schemas/types.ArtifactError'
          type: array
          uniqueItems: false
        logID:
          type: string
        snapID:
          type: string
        status:
          type: string
        warning:
          type: string
      required:
      - status
      type: object
...
    types.ChangePasswordRequest:
      properties:
        current:
          type: string
        new:
          minLength: 1
          type: string
      required:
      - current
      - new
      type: object
...
    types.CreateAPIKeyRequest:
      properties:
        name:
          maxLength: 128
          minLength: 1
          type: string
      required:
      - name
      type: object
    types.CreateAPIKeyResponse:
      properties:
        api_key:
          $ref: '#/components/schemas/types.APIKey'
        token:
          type: string
      required:
      - api_key
      - token
      type: object
...
    types.FriendInterface:
      properties:
        addresses:
          items:
            type: string
          type: array
          uniqueItems: false
        name:
          type: string
      required:
      - name
      type: object
    types.FriendPairRequest:
      properties:
        hostname:
          type: string
        identity_fingerprint:
          type: string
        identity_public_key:
          type: string
        port:
          type: integer
        scheme:
          enum:
          - http
          - https
          type: string
        signature:
          type: string
        x25519_public_key:
          type: string
      required:
      - hostname
      - identity_fingerprint
      - identity_public_key
      - signature
      type: object
    types.FriendPairResponse:
      properties:
        hostname:
          type: string
        identity_fingerprint:
          type: string
        identity_public_key:
          type: string
        paired:
          type: boolean
        x25519_public_key:
          type: string
      required:
      - hostname
      - identity_fingerprint
      - identity_public_key
      - paired
      type: object
    types.FriendPoll:
      properties:
        config:
          items:
            type: integer
          type: array
          uniqueItems: false
        config_hash:
          type: string
        exports:
          items:
            $ref: '#/components/schemas/types.FriendVar'
          type: array
          uniqueItems: false
        hello:
          $ref: '#/components/schemas/types.FriendsHello'
        interfaces:
          items:
            $ref: '#/components/schemas/types.FriendInterface'
          type: array
          uniqueItems: false
      required:
      - config_hash
      - hello
      type: object
...
    types.FriendVar:
      properties:
        name:
          type: string
        value:
          type: string
      required:
      - name
      - value
      type: object
    types.FriendsHello:
      properties:
        conntrackd:
          type: boolean
        conntrackd_running:
          type: boolean
        encryption:
          type: boolean
        hostname:
          type: string
        identity_fingerprint:
          type: string
        identity_public_key:
          type: string
        os:
          type: string
        signature:
          type: string
        uptime_sec:
          type: integer
        version:
          type: string
        vrrp:
          items:
            $ref: '#/components/schemas/types.FriendVRRPRole'
          type: array
          uniqueItems: false
        x25519_public_key:
          type: string
      required:
      - conntrackd
      - conntrackd_running
      - encryption
      - hostname
      - identity_fingerprint
      - identity_public_key
      - uptime_sec
      - version
      type: object
...
    types.Response-array_types_APIKey:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          items:
            $ref: '#/components/schemas/types.APIKey'
          type: array
          uniqueItems: false
        status:
          type: string
      type: object
...
    types.Response-array_types_FriendInterface:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          items:
            $ref: '#/components/schemas/types.FriendInterface'
          type: array
          uniqueItems: false
        status:
          type: string
      type: object
...
    types.Response-array_types_FriendVar:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          items:
            $ref: '#/components/schemas/types.FriendVar'
          type: array
          uniqueItems: false
        status:
          type: string
      type: object
...
    types.Response-config_Config:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/config.Config'
        status:
          type: string
      type: object
...
    types.Response-types_ApplyResult:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.ApplyResult'
        status:
          type: string
      type: object
...
    types.Response-types_CreateAPIKeyResponse:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.CreateAPIKeyResponse'
        status:
          type: string
      type: object
...
    types.Response-types_FriendPairResponse:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.FriendPairResponse'
        status:
          type: string
      type: object
    types.Response-types_FriendPoll:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.FriendPoll'
        status:
          type: string
      type: object
...
    types.Response-types_FriendsHello:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.FriendsHello'
        status:
          type: string
      type: object
...
    types.Response-types_StatusResponse:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.StatusResponse'
        status:
          type: string
      type: object
...
    types.StatusResponse:
      properties:
        status:
          type: string
      required:
      - status
      type: object
...
  securitySchemes:
    BearerAuth:
      description: '"Bearer <token>"'
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: Configuration and monitoring API for a Routier node.
  title: Routier API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/apply/confirm:
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_StatusResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: Confirm (keep) a pending apply
      tags:
      - apply
...
  /api/auth/api-keys:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-array_types_APIKey'
          description: OK
      security:
      - BearerAuth: []
      summary: List application API keys
      tags:
      - auth
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.CreateAPIKeyRequest'
              description: API key name
              summary: body
        description: API key name
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_CreateAPIKeyResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: Create an application API key
      tags:
      - auth
  /api/auth/api-keys/{id}:
    delete:
      parameters:
      - description: API key id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_StatusResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: Revoke an application API key
      tags:
      - auth
...
  /api/auth/password:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.ChangePasswordRequest'
              description: current and new password
              summary: body
        description: current and new password
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_StatusResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: Change the current user's password
      tags:
      - auth
...
  /api/config:
    get:
      parameters:
      - description: configuration layer
        in: query
        name: layer
        schema:
          enum:
          - advanced
          - simple
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-config_Config'
          description: OK
      security:
      - BearerAuth: []
      summary: Get the staged config
      tags:
      - config
...
  /api/config/apply:
    post:
      parameters:
      - description: configuration layer
        in: query
        name: layer
        schema:
          enum:
          - advanced
          - simple
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_ApplyResult'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_ApplyResult'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Apply the staged config
      tags:
      - config
...
  /api/config/import:
    put:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_StatusResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: Import a friend-pushed config payload
      tags:
      - config
...
  /api/config/raw:
    put:
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_StatusResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: Replace the whole staged config (raw YAML)
      tags:
      - config
...
  /api/friends/exports:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-array_types_FriendVar'
          description: OK
      security:
      - BearerAuth: []
      summary: nft variables this node publishes to its friends
      tags:
      - friends
  /api/friends/hello:
    get:
      parameters:
      - description: challenge nonce to sign
        in: query
        name: challenge
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_FriendsHello'
          description: OK
      security:
      - BearerAuth: []
      summary: Friend hello/handshake
      tags:
      - friends
  /api/friends/interfaces:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-array_types_FriendInterface'
          description: OK
      security:
      - BearerAuth: []
      summary: This node's interface addresses
      tags:
      - friends
  /api/friends/pair:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.FriendPairRequest'
              description: pairing payload
              summary: body
        description: pairing payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_FriendPairResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: Complete identity pairing (friend-token)
      tags:
      - friends
  /api/friends/poll:
    get:
      parameters:
      - description: config hash the caller already has
        in: query
        name: have
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_FriendPoll'
          description: OK
      security:
      - BearerAuth: []
      summary: Batched, shared-key-encrypted friend poll
      tags:
      - friends
...
  /api/v1/config:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OK
      security:
      - BearerAuth: []
      summary: Get the committed canonical configuration as YAML
      tags:
      - v1-config
...
servers:
- url: /